requirements.txt
discord.py[voice]
yt-dlp
PyNaCl
python-dotenv
davey